+++ packages/terra-core-docs/CHANGELOG.md
@@ -21,6 +21,7 @@
   * Added upgrade guide for `terra-action-header`.
   * Added Accessibility guide for `terra-button`.
   * Changes to Accessibility guide for `terra-button`.
+  * Added a11y changes to `terra-action-header` examples.
   * Added Accessibility guide for `terra-action-header`.
 
 * Fixed